Refactor ClusterStatus
