+++ src/ClusterManager/node_manager.py
@@ -24,7 +24,7 @@ from ResourceInfo import ResourceInfo
 
 import k8s_utils
 
-from resource_stat import Gpu
+from cluster_status import ClusterStatus
 
 k8s = k8s_utils.K8sUtil()
 
@@ -298,277 +298,6 @@ def _get_cluster_status():
     return cluster_status
 
 
-def str2bool(s):
-    return s.lower() in ["true", "1", "t", "y", "yes"]
-
-
-def get_node_statuses():
-    """Selects specific fields from Kubernetes node information.
-
-    Returns:
-        A dictionary of nodes with selected fields.
-    """
-    gpu_str = "nvidia.com/gpu"
-    nodes = k8s.get_all_nodes()
-
-    node_statuses = {}
-
-    for node in nodes:
-        # node is of class 'kubernetes.client.models.v1_node.V1Node'
-        name = node.metadata.name
-        labels = node.metadata.labels
-
-        gpu_type = ""
-        scheduled_service = []
-        if labels is not None:
-            for label, status in labels.items():
-                if status == "active" and label not in ["all", "default"]:
-                    scheduled_service.append(label)
-                if label == "gpuType":
-                    scheduled_service.append(status)
-                    gpu_type = status
-
-        allocatable = node.status.allocatable
-        gpu_allocatable = Gpu()
-        if allocatable is not None and gpu_str in allocatable:
-            gpu_num = int(allocatable[gpu_str])
-            gpu_allocatable = Gpu({gpu_type: gpu_num})
-
-        capacity = node.status.capacity
-        gpu_capacity = Gpu()
-        if capacity is not None and gpu_str in capacity:
-            gpu_num = int(capacity[gpu_str])
-            gpu_capacity = Gpu({gpu_type: gpu_num})
-
-        internal_ip = "unknown"
-
-        addresses = node.status.addresses
-        if addresses is not None:
-            for addr in addresses:
-                # addr is of class
-                # 'kubernetes.client.models.v1_node_address.V1NodeAddress'
-                if addr.type == "InternalIP":
-                    internal_ip = addr.address
-
-        unschedulable = node.spec.unschedulable
-        if unschedulable is not None and unschedulable is True:
-            unschedulable = True
-        else:
-            unschedulable = False
-
-        conditions = node.status.conditions
-        if conditions is not None:
-            for cond in conditions:
-                # cond is of class
-                # 'kubernetes.client.models.v1_node_condition.V1NodeCondition'
-                if cond.type == "Ready" and cond.status != "True":
-                    unschedulable = True
-
-        node_status = {
-            "name": name,
-            "labels": labels,
-            "gpuType": gpu_type,
-            "scheduled_service": scheduled_service,
-            "gpu_allocatable": gpu_allocatable,
-            "gpu_capacity": gpu_capacity,
-            "gpu_used": Gpu(),
-            "gpu_preemptable_used": Gpu(),
-            "InternalIP": internal_ip,
-            "pods": [],
-            "unschedulable": unschedulable
-        }
-
-        node_statuses[name] = node_status
-
-    return node_statuses
-
-
-def get_pod_statuses():
-    """Selects specific fields from Kubernetes pods information.
-
-    Returns:
-        A dictionary of pods with selected fields.
-    """
-    gpu_str = "nvidia.com/gpu"
-    pods = k8s.get_all_pods()
-
-    pod_statuses = {}
-    for pod in pods:
-        # pod is of class 'kubernetes.client.models.v1_pod.V1Pod'
-        phase = pod.status.phase
-        if phase in ["Succeeded", "Failed"]:
-            continue
-
-        name = pod.metadata.name
-        labels = pod.metadata.labels
-
-        gpu_type = ""
-        if labels is not None:
-            for label, status in labels.items():
-                if label == "gpuType":
-                    gpu_type = status
-
-        gpus = 0
-        preemptable_gpus = 0
-
-        username = None
-        if labels is not None and "userName" in labels:
-            username = labels.get("userName")
-
-        preemption_allowed = False
-        if labels is not None and "preemptionAllowed" in labels:
-            preemption_allowed = str2bool(labels["preemptionAllowed"])
-
-        node_name = pod.spec.node_name
-        pod_name = name
-        if username is not None:
-            pod_name += " : " + username
-
-        gpu_usage = get_job_gpu_usage(name)
-        if gpu_usage is not None:
-            pod_name += " (gpu usage:%s%%)" % gpu_usage
-            if gpu_usage <= 25:
-                pod_name += "!!!!!!"
-
-        containers = pod.spec.containers
-        if containers is not None:
-            for container in containers:
-                # container is of class
-                # 'kubernetes.client.models.v1_container.V1Container'
-                container_gpus = 0
-                # resources is of class
-                # 'kubernetes.client.models.v1_resource_requirements.V1ResourceRequirements'
-                resources = container.resources
-                requests = {}
-                if resources.requests is not None:
-                    requests = resources.requests
-
-                if gpu_str in requests:
-                    container_gpus = int(requests[gpu_str])
-
-                if preemption_allowed:
-                    preemptable_gpus += container_gpus
-                else:
-                    gpus += container_gpus
-
-                pod_name += " (gpu #:%s)" % container_gpus
-
-        pod_status = {
-            "pod_name": pod_name,
-            "node_name": node_name,
-            "username": username,
-            "gpus": Gpu({gpu_type: gpus}),
-            "preemptable_gpus": Gpu({gpu_type: preemptable_gpus}),
-            "gpuType": gpu_type
-        }
-        pod_statuses[name] = pod_status
-
-    return pod_statuses
-
-
-def update_node_statuses(node_statuses, pod_statuses):
-    for _, pod_status in pod_statuses.items():
-        pod_name = pod_status["pod_name"]
-        node_name = pod_status["node_name"]
-        pod_gpus = pod_status["gpus"]
-        pod_preemptable_gpus = pod_status["preemptable_gpus"]
-
-        if node_name not in node_statuses:
-            continue
-
-        # NOTE gpu_used may include those unallocatable gpus
-        node_status = node_statuses[node_name]
-        node_status["gpu_used"] += pod_gpus
-        node_status["gpu_preemptable_used"] += pod_preemptable_gpus
-        node_status["pods"].append(pod_name)
-
-
-def get_user_status(pod_statuses):
-    user_status = {}
-    user_status_preemptable = {}
-
-    for _, pod_status in pod_statuses.items():
-        username = pod_status["username"]
-        gpus = pod_status["gpus"]
-        preemptable_gpus = pod_status["preemptable_gpus"]
-        if username is not None:
-            if username not in user_status:
-                user_status[username] = Gpu()
-                user_status_preemptable[username] = Gpu()
-
-            user_status[username] += gpus
-            user_status_preemptable[username] += preemptable_gpus
-
-    return user_status, user_status_preemptable
-
-
-def compute_cluster_gpu_status(node_statuses):
-    capacity = Gpu()
-    used = Gpu()
-    avail = Gpu()
-    unschedulable = Gpu()
-    reserved = Gpu()
-
-    for node_name, node_status in node_statuses.items():
-        node_capacity = node_status["gpu_capacity"]
-        node_used = node_status["gpu_used"]
-        node_allocatable = node_status["gpu_allocatable"]
-        if node_status["unschedulable"]:
-            unschedulable += node_capacity
-            reserved += (node_capacity - node_used)
-        else:
-            # gpu_used may larger than allocatable: used one GPU that has
-            # uncorrectable errors
-            avail += (node_allocatable - node_used).min_zero()
-            unschedulable += (node_capacity - node_allocatable)
-            reserved += (node_capacity - node_allocatable)
-        used += node_used
-        capacity += node_capacity
-
-    logger.info("Cluster GPU status: capacity %s, used %s, avail %s, "
-                "unschedulable %s, reserved %s",
-                capacity, used, avail, unschedulable, reserved)
-
-    return capacity.resource, \
-        used.resource, \
-        avail.resource, \
-        unschedulable.resource, \
-        reserved.resource
-
-
-def compute_cluster_node_status(node_statuses):
-    for _, node_status in node_statuses.items():
-        gpu_capacity = node_status["gpu_capacity"].resource
-        gpu_used = node_status["gpu_used"].resource
-        gpu_preemptable_used = node_status["gpu_preemptable_used"].resource
-        gpu_allocatable = node_status["gpu_allocatable"].resource
-
-        node_status["gpu_capacity"] = gpu_capacity
-        node_status["gpu_used"] = gpu_used
-        node_status["gpu_preemptable_used"] = gpu_preemptable_used
-        node_status["gpu_allocatable"] = gpu_allocatable
-
-    return [node_status for _, node_status in node_statuses.items()]
-
-
-def compute_cluster_user_status(user_info, user_preemptable_info):
-    user_status = [
-        {
-            "userName": username,
-            "userGPU": user_gpu.resource
-        } for username, user_gpu in user_info.items()
-    ]
-
-    user_preemptable_status = [
-        {
-            "userName": username,
-            "userGPU": user_gpu.resource
-        } for username, user_gpu in user_preemptable_info.items()
-    ]
-
-    return user_status, user_preemptable_status
-
-
 def get_cluster_status():
     """Update in DB and returns cluster status.
 
@@ -577,40 +306,12 @@ def get_cluster_status():
     """
     cluster_status = {}
     try:
-        # Retrieve cluster information on nodes and pods
-        node_statuses = get_node_statuses()
-        pod_statuses = get_pod_statuses()
-        update_node_statuses(node_statuses, pod_statuses)
-        user_info, user_preemptable_info = get_user_status(pod_statuses)
-
-        # Compute GPU cluster status
-        gpu_capacity, gpu_used, gpu_avail, gpu_unschedulable, gpu_reserved = \
-            compute_cluster_gpu_status(node_statuses)
-
-        # TODO: Compute CPU cluster status
-
-        # TODO: Compute memory cluster status
-
-        # Compute cluster node status
-        node_status = compute_cluster_node_status(node_statuses)
-
-        # Compute cluster user status
-        user_status, user_status_preemptable = \
-            compute_cluster_user_status(user_info, user_preemptable_info)
-
-        # TODO: Deprecate typo "gpu_avaliable" in legacy code
-        cluster_status["gpu_avaliable"] = gpu_avail
-        cluster_status["gpu_available"] = gpu_avail
-        cluster_status["gpu_capacity"] = gpu_capacity
-        cluster_status["gpu_unschedulable"] = gpu_unschedulable
-        cluster_status["gpu_used"] = gpu_used
-        cluster_status["gpu_reserved"] = gpu_reserved
-        cluster_status["node_status"] = node_status
-        cluster_status["user_status"] = user_status
-        cluster_status["user_status_preemptable"] = user_status_preemptable
-
+        nodes = k8s.get_all_nodes()
+        pods = k8s.get_all_pods()
+        cs = ClusterStatus(config, nodes, pods)
+        cluster_status = cs.to_dict()
     except Exception:
-        logger.exception("Error in getting cluster status", exc_info=True)
+        logger.exception("Error in computing cluster status", exc_info=True)
 
     data_handler = None
     try:
