--- src/ClusterManager/test_node_manager.py
@@ -1,9 +1,0 @@
-#!/usr/bin/env python3
-
-from unittest import TestCase
-from node_manager import get_cluster_status
-
-
-class TestNodeManager(TestCase):
-    def test_get_cluster_status(self):
-        pass
